Add rate limiting on BPF events map
This change introduces a rate limiter that prevents from high CPU
utilization by cilium-agent. This is particuraly important on smaller
VMs where consuming ~3vCPU can starve other proceses on the node. In the
same time CPU limit on the whole pod is not a good solution, as it may
affect network programing latency.

The rate limiter operates on BPF level, preventing an event to be
generated and written to BPF events map. It utilizes the rate limiter
already used for ICMPv6 messages, which implementation is based on a
token system. The rate limiter is configured with 2 values - rate limit
(how many tokens are refreshed every second) and burst limit (the
maximum number of tokens).

Option should be used when you either need to limit cilium-agent CPU
usage (without disabling hubble subsystem at all) or you can observe a
lot of event lost in the observer queue (as this is wasting CPU time on
processing events that won't be consumed anyway.

This is a follow-up to https://github.com/cilium/cilium/pull/25385,
which implemented similar behaviour, but in user space. This
implementation provides rate limiting of events at the earliest stage of
event lifecycle, preventing the memory allocation for a message.

Signed-off-by: Michal Siwinski <siwy@google.com>